requirements_cpu.txt
torch>=1.10.0
torchvision>=0.11.0
numpy==1.22.3
scikit-image==0.16.2
scipy==1.7.2
Pillow==8.4.0
matplotlib>=3.2.2
pyFFTW==0.13.1
